Fixing DB ping interval logic.
The DB ping interval logic was wrong as it was not initializing the DB Pinger.
On MySQL Sessions through Bastion this caused the session to get disconnected
after a few minutes.

This patch fixes the problem by correctly instantiating the DB Pinger, also
configured a default timeout of 60 second for MySQL Sessions through Bastion
to ensure the connection is kept open if not used.

Change-Id: I1396a9350d9e6096880f1c9e8e9dcdcb8b5de91f