Restructure MySQL keyword handling and updated them
MySQL keywords are held in a JSON file and are updated manually from the MySQL server grammar. Previously they were in a single big list, making it difficult to update them. Now there's a list for each server version (just like in the server grammar, see `%token` entries). This simplifies handling as only the last list is updated.

Additionally, the code editor was changed to use this updated list of keywords, instead of the hard coded one in the Monaco MySQL language contribution.

Change-Id: I78252f504da4307ed5799bd649b29fb027c2a328