Fixed active item tracking in document selector
In embedded scenarios the document selector shows not only documents from one connection, but a tree of connections with their documents. This means that fixed components like admin pages need constant, but per-connection ids. Additionally, the actual selection is determined by looking at the active element per connection. With multiple connections the item in the last connection was always selected. This is now fixed.

Change-Id: I5f77f15f7cb95d88c7389e7a1e6d9e606c0f6856