Fix usage of MySQL Shell sessions on external plugins.
Plugin functions that work with a standard Shell Session, get the
"session" parameter resolved to the raw MySQL Session contained within
the DbMySQLSession object on the web layer.

Since the functions called from the web interface are now being handled
through the RequestHandler which executes in a separate thread, if more
than one function requiring a session is called at the same time, a race
condition was hit causing the Shell Process to die. This patch fixes the
problem by:

- Enhancing the DbSession class hierarchy to support the standard method
  for SQL execution in the Shell Sessions: run_sql(sql[, args]).
- Adding a locking mechanism to avoid the session from being used at the
  same time i.e. by two functions being executed in its own
  RequestHandler.

Since the MySQL Shell did not allow passing python objects as
parameters, a change had to be done on the Shell itself to no longer
restrict this.

So now the plugin functions written for standard Shell sessions will
support execution either with a standard MySQL Shell Session, or a
DbSession object from the web layer.

Change-Id: I4e140f8697e6e6c511ab405e9fd4aaf86611466a