WL #15410 Rework of the Event Handling
The communication of the MSG frontend with its backend requires sending JSON objects over a web socket connection back and forth. The format of these objects are determined by the BE and the FE has to follow that.

This format has changed over time and new features have been introduced that need extended handling. Additionally, the way events are handled does not allow to fully use TS async methodology, leading partially to highly convoluted code.

This WL aims to rework that handling in a way that nicely integrates with async functionality for much easier consumption by the app in general.

Change-Id: I791ea8573ecf04e52cf30d5cddac77fbd12a4038