Bug#35135302 - MRS - Custom URL for Access Token field is not saved
The main issue was that we were not allowing None to be set for the items_per_page.

The solution was to update the protocol files along with the handling of the value(starting in the dialog response, ending in the backend)

Also important tests were added for the authentication vendors and authentication apps.

Change-Id: I6f6b2916d8e91b109bf49f65f2801397b09b3ad3