BUG#35341478 Display error line and column in DB Notebook
A previous commit introduced a regression that caused the DB Notebook in
the MySQL Shell GUI Extension for VS Code to stop displaying the error
line and column numbers for incorrect syntax and other errors in code
written in one of the supported languages.

The issue is caused by a nonintentional change to the regular expression
used to parse the raw error stack trace. This patch reverts that
specific change in the regular expression.

Change-Id: I42b02c059a1fc203cd2c03dbe0aabe36cd694432