WL#15716 Allow reading related data across tables
With JSON/Relational duality, one can expand foreign key relationships
between different database objects. To avoid excessive details to be
included in the result set, the MRS SDK should allow to pick which
fields that span different tables should be included or excluded in the
result set.

This patch introduces the changes required to allow the MRS SDK to read
related data across tables and include or exclude it from the result set
alongside the proper type-safety and auto-completion infrastructure.

Change-Id: Ib5665c2b54b8c876d9accba0df0d4527bbef3e9a