Reorganisation of the MRS dialogs
The dialogs to add/edit MRS objects had originally been added to the extension code. However, these dialogs (being already to a large extension frontend code) are needed in the frontend too, for handling the same MRS functionality like for the extension. For this reason all related code of these dialogs has been moved to the FE. This required a rework of dialog handling to allow waiting for a dialog to be closed. The original requisition based handling of all other dialogs remains unchanged. This patch contains these changes:

- Split of the DialogType enum into 3, to account for normal dialogs, MDS dialogs and MRS dialogs.
- Renamed the MrsTurnstile component to MrsHub.
- Simplification of the command handler setup in the extension.
- Removal of the unused MDSModule component.
- Removal of all MRS dialogs from the DialogHost component. They were moved to the MrsHub component.
- Added a notification to modify the VS Code status bar in the extension, from the frontend.
- Added a parameter to the Semaphore class, which allows to specify a notification reason/result.
- Moved string related helper functions from helpers.ts to string-helpers.ts.
- Fixed unit tests and added new tests for the new code.

Change-Id: I6252cca05e5a682fd347ee4b356bcc56ca3f2941