BUG#36798984 Connection through bastion fails after expired/deleted bastion session
The logic to renew the bastion session was implemented only for reconnections, but
it never considered the same scenario was valid for a first time connection. This
patch refactors the handling of the SessionSetupTasks in a way they are properly
setup and cleaned up before a connection attempt is made, in all scenarios.

Change-Id: Ia087f64d037f771d6343fed8fa4f5fe7980da032