Show the correct line for syntax errors
For syntax errors the server sends back a line number with the error message (and only for those). That line number is local to the given query, but should show correspond to the correct line number within the executing block and also account for leading line breaks, which the server ignores when computing the error line.

Change-Id: I8da97a73dba59bf9b64b4252d808640aa029b012