Disable new sample app directory for Jest extension
The "MRS Notes" sample app directory and the corresponding reference
folder in the VSCode workspace have changed in d7f85f1f. The workspace
was also previously configured to prevent the Jest extension from
running tests on the old folder and, in this case, the configuration was
not updated. This means the Jest extension is now trying to run tests on
the new/renamed folder, which is not supposed to happen.

This patch updates the "MRS Notes" sample app workspace folder in the
list of folders disabled by the Jest extension.

Change-Id: Ic3befa2464f33ed26c8bccc32937d65d9494e5cc