Changed clipboard handling in the extension
The previous solution did not work for cut and copy clipboard actions (in code editors) and was focused on getting key presses for clipboard actions forwarded, which is the wrong way as that does not allow accessing the clipboard in embedded scenarios. Instead we can enable clipboard access in the iframe and register hotkeys in embedded mode (in the code editor) to do the clipboard actions. This also allows to remove the paste action handler code.

Also removed an unused field in the code editor model.

Change-Id: I3e739c2cb5f67d7eb24c4ec865deb01abc3a4a80