BUG#36027690 Fix REST Object URL opened by web browser
Trying to open a REST Object URL with the Web Browser using the
corresponding feature available in the MySQL Shell GUI Extension for VS
Code yields a connection error.

This is because the generated URL uses the default port for the MySQL
Router HTTP server (8443) and now, each connection gets assigned a
different port for the REST service endpoint.

This patch introduces the changes to ensure the Web Browser opens a REST
Object URL generated using the correct port of the REST service endpoint.

Change-Id: I466e87ce88707fab3fffabea02cbdf57e7518f44