E2E extension tests new function
This patch creates the function 'removeAllDatabaseConnections()'.
This function is now not only called at the end of each test suite, but also at the beginning of each test suite. It has happened that because of buggy tests, the DB connections are not removed at the end of the tests, causing test failures on further executions. With this patch, we ensure there are no DB connections at the beginning of the tests

Change-Id: Idc1bada93a288e6749292e539231b2e67d2c0c33