WL#15718 Application-level interfaces for the MRS SDK
A response sent by the MySQL Router to an HTTP request in scope of the
MySQL REST service includes the (selected) database object fields and
values alongside additional hypermedia properties and resource metadata
which can be used for navigation between resources, pagination and
caching.

These properties are usefull from the standpoint of a stateless HTTP
client such as a Web Browser (HATEOAS) but not so much for a stateful
application using the MRS SDK. In that case, they can even increase
bloat, for instance when performing a resource update directly using an
object instance previously retrieved with the REST API, wherein some of
those properties are useless. Ultimately, hypermedia properties can be
leveraged by the SDK, in certain circumstances, but not the application
itself.

Additionally, by dropping these properties, we can have a scenario where
the database object types can be directly shared between the client and
the server in the MRS SDK.

This patch introduces some changes to simplify the resource details at
the application level by hiding hypermedia properties under a proper
abstraction and sharing database object types between client and server.

Change-Id: If0a64c29c1174ecc671f0c0f50aa61b25e88d1f0