Change setup of extension E2e tests
We are seeing the extension e2e tests job to hang sometimes, on the extension installation. This step is done as a background job, to be run in parallel for all 6 suites and this process is causing issues. The new setup installs the extension only on 1 vscode instance, and then copy the extracted contents of the vsix file into the other test suite folders.

Change-Id: I6d515e889021331a02e4b6124d594ce50efa9aa0