BUG#36363355 Connect with scheme and custom port/socket
In a Shell session created with the MySQL Shell GUI, it is possible to
connect to the MySQL server using the "\connect" command whilst
specifying a connection string.

However, for the first connection attempt, if the connection string
contains a scheme such as "mysql://" or "mysqlx://", and uses a
non-standard port or Unix socket file location, the connection will fail
after setting the password, and the UI will display an error message.

This patch introduces the changes required to allow connections under
such circumstances.

Change-Id: I2331752318e60a3505bc241d58d22fa6df2a85ee