WL#16158 MRS SDK API improvements to handle multiple items
With WL#15718, an application using the MRS TypeScript SDK is no longer
working with with the raw JSON response sent by the MySQL Router. Items
retrieved from a database object are organized in a typed client-side
array structure instead which does not contain hypermedia properties
such as "hasMore", which is particularly relevant for pagination.

From the application standpoint, looping through the items with offset
pagination can be done by skipping the number of pages multiplied by the
number of results per page whilst the SDK can keep additional state to
ensure extraneous requests are sent to the MySQL Router.

Offset pagination does not scale at a database level, because, if one
skips thousands of records and only take the first few dozen, the
database still has to traverse those first thousands before reaching
and returning the ones that are expected.

An alternative is to a cursor and a page size to return a limited set of
results before or after a given cursor. The cursor bookmarks the client
location in the result set and must be a unique, sequential column such
as an "AUTO_INCREMENT" or "TIMESTAMP" column.

This patch introduces improvements for the MRS TypeScript SDK to account
for a multitude of scenarios, such as those described above, that arise
when retrieving and handling multiple resources from a REST service.

Change-Id: I5c8658465f4c5a960645fb175f39461636346d93