Remove plain HTTP API from the TypeScript SDK
The MRS SDK settled on a Prisma-inspired API that uses CRUD-like
commands to perform operations over the resources in a REST service. The
TypeScript SDK still offers a plain HTTP wrapper API on which the CRUD
commands are built on top of.

However, due to the fluent nature of this HTTP API, it ends up causing
more issues than it solves when it comes to account for internal state
when building a request. Thus, a decision was made to remove the API
altogether.

This patch removes said API and simplifies some internal workflows to
build the underlying HTTP requests.

Change-Id: I8f4378774eab0ed7245a0336ab9f27dfdd353b39