BUG#36589986 Import all the MRS SDK required datatypes
If a REST service is composed of database objects that only contain
fields whose column datatype can be directly mapped to a primitive
TypeScript type and one or more of those fields are nullable, the
generated SDK files will not include the "MaybeNull" custom type, which
is required for handling operations with those nullable fields.

The issue is due to an off-by-one error in the codegen service in the
GUI backend plugin, which only accounts for more than one required
datatype (instead of one or more).

This patch introduces the changes required to address the issue and
ensures the resulting import statement contains all the required
datatypes.

Change-Id: I57b4ade4b5ede30003052a968e3349468de3b0d2