bugfix: avoid setting focus on null element in the sample app
When accepting a shared note with an invitation code in the MRS sample
app, the app tries to set the focus on the text area that displays the
contents of the selected note. However, when the list of notes is empty,
there is no note to begin with. This means no element has been rendered
and the "current" property in the element reference created with
Preact's "createRef()" has not yet been assigned and is null.

This patch removes the extraneous function call that attempts to set the
focus on the non-existing element.

Change-Id: I23a84d99670c1ba9e717cae01ef4a6711835d2b1