BUG#36445530 Restrict findUnique filter to unique fields
The findUnique command in the MRS TypeScript SDK should only be able to
work with a filter that narrows the search to a single item. Since the
search is performed using a unique field, which by itself should be
enough to identify that item, logical operators such as AND and OR are
extraneous and should not be possible to use them in the query filter.

This patch introduces the changes required to restrict the query filter
to the set of unique fields of a given database object.

Change-Id: Iac3ec994b40c5918deb5d0e5bccd59f0ac1b9fee