WL#16469 Read-your-write consistency for the TypeScript SDK
With multiple MySQL server instances running as an InnoDB
Cluster/ClusterSet, data read from one instance might be dependent on
data written on a different instance, which might have not been yet
replicated to the instance where data is being read from.

To solve this issue, MySQL uses a Global Transaction Identifier (GTID)
which allows to reconcile different transactions in the cluster. When
this functionality is enabled, from the MRS standpoint, the MySQL Router
sends back the GTID value for each operation that modifies data (DELETE,
POST and PUT). The client then sends back that GTID for operations that
require consistent reads. This can be costly, so it must be explicitly
enabled by the client.

This patch introduces the changes required to introduce a boolean toggle
("readOwnWrites") for "find*" and "delete*" commands available in the
MRS TypeScript SDK. This toggle can be used by an application, running
on top of a MySQL instance cluster, that requires read consistency for
those kind of operations.

Change-Id: I65d4cf3bd27d5b94df8eff2750800d73f68c0d33